Added assertions to enable numeric values to be validated as being positive, negative, or nonNegative for OVPMS-584 Charges and payments permit entry of negative values which cause allocation and balance miscalculations
svn path=/framework/trunk/; revision=2020

+++ openvpms-framework/test/java/org/openvpms/component/business/service/archetype/descriptor/cache/ArchetypeDescriptorCacheFSTestCase.java
@@ -171,9 +171,8 @@ public class ArchetypeDescriptorCacheFSTestCase extends BaseTestCase {
         String dir = (String) params.get("dir");
         String assertionFile = (String) params.get("assertionFile");
         String extension = (String) params.get("extension");
-        int recordCount1 = ((Integer) this.getTestData().getTestCaseParameter(
-                "testLoadingArchetypesFromDir", "normal", "recordCount1"))
-                .intValue(); 
+        int recordCount1 = (Integer) this.getTestData().getTestCaseParameter(
+                "testLoadingArchetypesFromDir", "normal", "recordCount1");
 
         ArchetypeDescriptorCacheFS cache = new ArchetypeDescriptorCacheFS(dir,
                 new String[] {extension}, assertionFile);
